+++ test/test_helper.rb
@@ -37,7 +37,7 @@ def mock_list_models_response_payload
           ]
         },
         {
-          "id": "mistral-small",
+          "id": "mistral-small-latest",
           "object": "model",
           "created": 1703186988,
           "owned_by": "mistralai",
@@ -145,7 +145,7 @@ def mock_chat_response_payload
           "index": 0
         }
       ],
-      "model": "mistral-small",
+      "model": "mistral-small-latest",
       "usage": {"prompt_tokens": 90, "total_tokens": 90, "completion_tokens": 0}
     }
   JSON
@@ -155,7 +155,7 @@ def mock_chat_response_streaming_payload
   [
     'data: ' + {
       'id' => 'cmpl-8cd9019d21ba490aa6b9740f5d0a883e',
-      'model' => 'mistral-small',
+      'model' => 'mistral-small-latest',
       'choices' => [
         {
           'index' => 0,
@@ -169,7 +169,7 @@ def mock_chat_response_streaming_payload
         'id' => 'cmpl-8cd9019d21ba490aa6b9740f5d0a883e',
         'object' => 'chat.completion.chunk',
         'created' => 1_703_168_544,
-        'model' => 'mistral-small',
+        'model' => 'mistral-small-latest',
         'choices' => [
           {
             'index' => i,
@@ -181,4 +181,28 @@ def mock_chat_response_streaming_payload
     end,
     "data: [DONE]\n\n"
   ]
+end
+
+def mock_completion_response_payload
+  {
+    'id' => 'chat-98c8c60e3fbf4fc49658eddaf447357c',
+    'object' => 'chat.completion',
+    'created' => 1_703_165_682,
+    'choices' => [
+      {
+        'finish_reason' => 'stop',
+        'message' => {
+          'role' => 'assistant',
+          'content' => ' a + b'
+        },
+        'index' => 0
+      }
+    ],
+    'model' => 'mistral-small-latest',
+    'usage' => {
+      'prompt_tokens' => 90,
+      'total_tokens' => 90,
+      'completion_tokens' => 0
+    }
+  }.to_json
 end