+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.1/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-05-30
+
+### Added
+
+- Added support for completion requests which you can use to query their latest model
+[codestral](https://mistral.ai/news/codestral/).
+See [this example](https://github.com/wilsonsilva/mistral/blob/0.3.0/examples/code_completion.rb) to get started.
+
 ## [0.2.0] - 2024-05-23
 
 ### Added
@@ -26,5 +34,6 @@ Ports [mistralai/client-python#86](https://github.com/mistralai/client-python/pu
 - Initial release. Feature parity with `v0.1.8` of the
 [mistralai/client-python](https://github.com/mistralai/client-python)
 
+[0.3.0]: https://github.com/wilsonsilva/mistral/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/wilsonsilva/mistral/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/wilsonsilva/mistral/compare/28e7c9...v0.1.0